- fix a deadlock - improve client stats - core compatibility fixes
git-svn-id: http://svn.airdcpp.net/airsvn/trunk@1710 ec3e54a6-43b8-4a3e-9c49-2a733a73d9fe

+++ UserQueue.cpp
@@ -114,7 +114,7 @@ QueueItemPtr UserQueue::getNextBundleQI(const UserPtr& aUser, const StringSet& r
 	if(i != userBundleQueue.end()) {
 		dcassert(!i->second.empty());
 		for (auto& b: i->second) {
-			if (bundleLimit > 0 && runningBundles.size() >=  bundleLimit && runningBundles.find(b->getToken()) == runningBundles.end())
+			if (bundleLimit > 0 && static_cast<int>(runningBundles.size()) >=  bundleLimit && runningBundles.find(b->getToken()) == runningBundles.end())
 				continue;
 
 			if (b->getPriority() < minPrio) {
